feat(listing): live listings section for sidebar
